requirements.txt
alabaster==0.7.13
annotated-types==0.5.0
anyio==3.7.1
Babel==2.12.1
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
confection==0.1.0
courlan==0.9.3
cymem==2.0.7
dateparser==1.1.8
docopt==0.6.2
docutils==0.20.1
fasttext==0.9.2
fasttext-langdetect==1.0.5
gensim==4.3.1
h11==0.14.0
htmldate==1.4.3
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
jusText==3.0.0
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
murmurhash==1.0.9
numpy==1.25.1
packaging==23.1
pathy==0.10.2
preshed==3.0.8
pybind11==2.10.4
pydantic==1.10.7
pydantic_core==2.3.0
Pygments==2.15.1
python-dateutil==2.8.2
python-telegram-bot[job-queue]==20.4
pytz==2023.3
regex==2023.6.3
requests==2.31.0
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==6.2.1
sphinx-rtd-theme==1.2.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.6
thinc==8.1.10
tld==0.13
tqdm==4.65.0
trafilatura==1.6.1
typer==0.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.3
warcio==1.7.4
wasabi==1.1.2
wget==3.2
yarg==0.1.9
